dev/requirements.txt
ambient_wx==0.1.2
apscheduler==3.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9